refactor(language-service): clean up and exports and consolidate types
