+++ packages/language-service/language-service.ts
@@ -15,6 +15,5 @@
  */
 export {createLanguageService} from './src/language_service';
 export * from './src/ts_plugin';
-export {Completion, Completions, Declaration, Declarations, Definition, Diagnostic, Diagnostics, Hover, HoverTextSection, LanguageService, LanguageServiceHost, Location, Span, TemplateSource, TemplateSources} from './src/types';
+export {Declaration, Definition, Diagnostic, LanguageService, LanguageServiceHost, Span, TemplateSource} from './src/types';
 export {TypeScriptServiceHost, createLanguageServiceFromTypescript} from './src/typescript_host';
-export {VERSION} from './src/version';
